requirements.txt
openai
python-dotenv
tiktoken
scipy
gitpython